+++ CHANGELOG.md
@@ -16,6 +16,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Adds a gh action that checks that the changelog is updated
 - Adds a gh action that deploys new releases automatically to pypi
 - Warn users if case default panels are outdated
+- Define institute-specific gene panels for filtering in institute settings
+- Use institute-specific gene panels in variants filtering
 
 ### Fixed
 - Report pages redirect to login instead of crashing when session expires